requirements.txt
Flask==2.3.3
twilio==8.5.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
openai==0.28.0
python-dotenv==1.0.0
requests==2.31.0
pyttsx3==2.90
SpeechRecognition==3.10.0
numpy==1.24.3
python-jose==3.3.0
pydub==0.25.1
